feat: widgets 폴더 내 times 관련 코드와 타입을 entities 폴더로 이동
