+++ src/entities/time/model/types.ts
@@ -1,0 +1,7 @@
+export type Time = {
+  startTime: string;
+  endTime: string | null;
+  ranges: number[];
+  isNextDay?: boolean;
+  isDefault?: boolean;
+};
