+++ src/widgets/time/index.ts
@@ -1,6 +1,4 @@
-export { getDefaultTimes } from './lib/getDefaultTimes';
-
 export { TimeBottomSheet } from './ui/TimeBottomSheet/TimeBottomSheet';
 
 export { TimeSelector } from './ui/TimeSelector/TimeSelector';
-export type { Time, Day } from './model/types';
+export type { Day } from './model/types';
