+++ src/widgets/time/model/types.ts
@@ -1,10 +1,3 @@
-export type Time = {
-  startTime: string;
-  endTime: string | null;
-  ranges: number[];
-  isNextDay?: boolean;
-  isDefault?: boolean;
-};
 export type DraggingRangeStatus = 'currentDay' | 'nextDay' | 'impossible';
 export type Day = '오늘' | '내일' | '모레';
